requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
Pillow==9.1.0
scikit_learn==1.0.2
